feat: Finish fragment
