+++ src/components/Day/components/index.tsx
@@ -1,0 +1,3 @@
+export { default as DayMovie } from "./DayMovie";
+export { default as DayTV } from "./DayTV";
+export { default as DaySpotify } from "./DaySpotify";
