+++ src/components/Day/index.tsx
@@ -1,0 +1,1 @@
+export { default as Day } from "./Day";
