+++ src/components/Edit/components/index.tsx
@@ -1,0 +1,3 @@
+export { default as EditMovie } from "./EditMovie";
+export { default as EditTV } from "./EditTV";
+export { default as EditSpotify } from "./EditSpotify";
