+++ src/components/Edit/index.tsx
@@ -1,0 +1,1 @@
+export { default as Edit } from "./Edit";
