+++ src/components/selected/components/index.tsx
@@ -1,0 +1,3 @@
+export { default as SelectedMovie } from "./SelectedMovie";
+export { default as SelectedTV } from "./SelectedTV";
+export { default as SelectedSpotify } from "./SelectedSpotify";
